+++ include/build_info.h
@@ -1,4 +1,4 @@
 #pragma once
 
 constexpr const char* BUILD_ID =
-    "7IN-20260801-PRODUCT56-R6-SYSTEM-UI";
+    "7IN-20260801-PRODUCT57-AIRPORT-DATABASE-SETUP";